requirements.txt
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
open3d>=0.17.0
Pillow>=9.0.0
timm>=0.9.0
